584: some determination fixes
