+++ src/Hedwig/ClientApp/src/utils/models/familyDetermination.ts
@@ -15,3 +15,21 @@ export const familyDeterminationNotDisclosed = (enrollment: Enrollment) : boolea
 
 	return determinations[0].notDisclosed;
 };
+
+export function determinationArgsAreValid(args: any) {
+  if (args.notDisclosed) {
+    args.numberOfPeople = null;
+    args.income = null;
+    args.determinationDate = null;
+    return true;
+  }
+
+	if(args.numberOfPeople 
+		&& args.income !== undefined && args.income !== null
+		&& args.determinationDate) {
+    args.notDisclosed = false;
+    return true;
+  }
+
+  return false;
+}
