+++ src/Hedwig/ClientApp/src/utils/models/index.ts
@@ -1,6 +1,6 @@
 export * from './gender';
 export * from './child';
 export * from './family';
-export * from './determination';
+export * from './familyDetermination';
 export * from './funding';
 export * from './enrollment';
